RA-1533 Update the registration module to permit editing all the sections
This commit updates RegisterPatientFragmentController, and adds a new method that handles the edit of the entire patient registration form (without modifying the existing code that handles adding a new registration). When an extra query parameter (patientIdNumber) is included, the controller edits the existing patient. This solution is backward compatible: it doesn't alter the current patient registration flow and adds the possibility to also edit a patient